Updated to use stored AWS credentials so secret key no longer needs to be in the MSBuild file.
Updated S3BuildPublisher to accept a MSBuild ItemList of FileNames (@() style) and to include a public accessible flag to make the files stored public.

Updated S3Helper to publish the files and use constructor for passing of credentials and instantiation of the S3 client.

Added ReadMe.txt file to include details on the project, debug info etc.

git-svn-id: https://snowcode.googlecode.com/svn/trunk/2010-04-TramDepot/Snowcode.S3BuildPublisher@83 eb82f096-58cc-11de-8832-3f4c773c73ae